Chore: TypeScript migration - move `notifications/push` to `lib/notifications` (#3999)
* chore: move `notifications/push` to `lib/notifications/push`

* update: move `lib/notifications/push` to `lib/notifications`

* remove: push.ios.ts
